requirements.txt
defusedxml==0.6.0
diff-match-patch==20200713
Django==2.2
django-import-export==2.5.0
django-userforeignkey==0.3.0
et-xmlfile==1.0.1
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.0
openpyxl==3.0.2
Pillow==8.0.1
python-barcode==0.13.1
pytz==2020.4
PyYAML==5.3
reportlab==3.6.1
sqlparse==0.4.1
tablib==0.14.0
xlrd==1.2.0
xlwt==1.3.0